+++ releases.md
@@ -19,11 +19,14 @@ Changes are grouped as follows:
 
 - Geo-location attribute and resource type.
 
+
 ## [1.8.0-SNAPSHOT]
 
 ### Added
 
 - Experimental streaming support for `events` and `assets`.
+
+- Support for 3D Models (2021-12-06)
 
 ## [1.7.1] 2021-11-23